Finish UI based game functionality
At this point, this is a working game. There is no AI that will
play against a single person, it is designed as a 2 person game.
As of now, the 2 players cannot choose their marks (player1 is x and player2 is o).
There is some refactoring to be done as several html elements were defined as variables
within multiple functions, when it appears that they would be better off as global variables.